refactor(FIN-033): standardize import paths for constants and enhance user and currency models
