+++ src/common/constants/currency-requirements.constant.ts
