+++ src/common/enums/support-languages.enum.ts
@@ -1,4 +1,6 @@
-export const enum SupportLanguages {
-  Ukrainian = 'uk',
-  English = 'en',
-}
+export const SupportLanguages = {
+  Ukrainian: 'uk',
+  English: 'en',
+};
+
+export type SupportLanguages = (typeof SupportLanguages)[keyof typeof SupportLanguages];
