+++ src/server/entities/country/infrastructure/country.orm.ts
@@ -1,8 +1,8 @@
 import { Column, Entity, OneToMany } from 'typeorm';
 import { DefaultTableColumnsOrm } from '../../../shared/infrastructure/default-table-columns.orm';
 import { type CountryAndLocale } from '@common/records/countries.record';
-import { CountryRequirementsConstant } from '@common/domains/lookups/constants/country-requirements.constant';
 import { UserOrm } from '@backend/entities/user/infrastructure/user.orm';
+import { CountryRequirementsConstant } from '@common/constants/country-requirements.constant';
 
 @Entity('country')
 export class CountryOrm extends DefaultTableColumnsOrm implements CountryAndLocale {
