+++ src/server/entities/user/infrastructure/user.orm.ts
@@ -1,11 +1,14 @@
-import { Column, Entity, JoinColumn, ManyToOne, OneToMany, RelationId } from 'typeorm';
+import { Column, Entity, JoinColumn, ManyToOne, OneToMany } from 'typeorm';
 import { DefaultTableColumnsOrm } from '../../../shared/infrastructure/default-table-columns.orm';
 import { type FullUserData } from '@common/records/user.record';
-import { UserRequirements } from '@common/domains/user/constants/user-requirements.constant';
 import { RoleEnum } from '@common/domains/user/enums/role.enum';
 import { type RegularEntryOrm } from '@backend/entities/regular-entry/infrastructure/regular-entry.orm';
 import { CountryOrm } from '@backend/entities/country/infrastructure/country.orm';
 import { CurrencyOrm } from '@backend/entities/currency/infrastructure/currency.orm';
+import { UserRequirements } from '@common/constants/user-requirements.constant';
+import { CurrencyRequirements } from '@common/constants/currency-requirements.constant';
+import { CountryRequirementsConstant } from '@common/constants/country-requirements.constant';
+import { SupportLanguages } from '@common/enums/support-languages.enum';
 
 @Entity('user')
 export class UserOrm extends DefaultTableColumnsOrm implements FullUserData {
@@ -23,12 +26,12 @@ export class UserOrm extends DefaultTableColumnsOrm implements FullUserData {
 
   @ManyToOne(() => CurrencyOrm, (currency) => currency.users)
   @JoinColumn({
-    name: 'currency_code',
-    referencedColumnName: 'code',
+    name: 'currencyCode',
+    referencedColumnName: 'currencyCode',
   })
   currency!: CurrencyOrm;
 
-  @RelationId((user: UserOrm) => user.currency)
+  @Column({ type: 'varchar', length: CurrencyRequirements.MaxCurrencyCodeLength })
   currencyCode!: string;
 
   @ManyToOne(() => CountryOrm, (country) => country.users)
@@ -38,8 +41,11 @@ export class UserOrm extends DefaultTableColumnsOrm implements FullUserData {
   })
   country!: CountryOrm;
 
-  @RelationId((user: UserOrm) => user.country)
+  @Column({ type: 'varchar', length: CountryRequirementsConstant.MaxLocaleLength })
   locale!: string;
+
+  @Column({ type: 'enum', enum: Object.values(SupportLanguages), default: SupportLanguages.Ukrainian })
+  language!: SupportLanguages;
 
   @OneToMany('RegularEntryOrm', 'user')
   regularEntries!: RegularEntryOrm[];
